Add torch model & reduce the size of the torch hitbox - Pull out the block model selection from ChunkModel into BlockRenderer - Use int[]'s for texture coordinates, only reducing to shorts at the end - Add offset when setting position of an AABBCollider - Change the selection highlight back to a black box, offset to make   more visible

+++ src/ddb/io/voxelnet/client/render/FontRenderer.java
@@ -73,7 +73,7 @@ public class FontRenderer
 		float yMin = (float)Math.round((y+ 0f*FontScale));
 		float yMax = (float)Math.round((y+16f*FontScale));
 		
-		short[] texCoords = fontAtlas.getPixelPositions(chr);
+		int[] texCoords = fontAtlas.getPixelPositions(chr);
 		// Add the quad
 		// 0-3
 		// |\|
